pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mpdex" 
version = "0.1.0"
description = "Multi Perp DEX base class and exchange wrappers"
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
  "aiohttp>=3.10",
  "pynacl>=1.5",
  "ccxt>=4.4",     
  "eth-hash==0.7.1",
  "curl-cffi>=0.13.0",
  "solders>=0.19.0",
  "base58>=2.1.1",
  #"lighter-sdk @ git+https://github.com/elliottech/lighter-python.git ; sys_platform != 'win32'",
  # due to spot update, lighter-sdk needs to be updated. do it later  # ; sys_platform == 'win32'",
  "lighter-sdk @ git+https://github.com/elliottech/lighter-python.git@c9b8de6fd6864ede0fd7fb61d85ffb3614d178c9",
  "grvt-pysdk>=0.1.24",
  "cairo-lang==0.13.5",
  "msgpack==1.1.2",

  # [ADDED] Windows(Py3.10, AMD64) 한정: fastecdsa 사전빌드 휠 자동 설치
  # - pip, uv 모두 PEP 508 마커를 해석합니다.
  # - 링크는 사용자가 제공한 공식 PyPI CDN(files.pythonhosted.org) 휠입니다.
  "fastecdsa @ https://files.pythonhosted.org/packages/88/e4/0081bef22304409919e27e092b84f140e458901239ff9501f330cf593f0d/fastecdsa-2.3.0-cp310-cp310-win_amd64.whl ; python_version == '3.10' and platform_system == 'Windows' and platform_machine == 'AMD64'"
]


[tool.setuptools]
# 설치 대상 패키지를 명시적으로 제한하여 keys/, test_exchanges/는 설치 제외
packages = ["mpdex", "wrappers", "mpdex.utils"]
# 루트에 있는 단일 모듈도 함께 배포
py-modules = ["multi_perp_dex", "exchange_factory"]

[tool.setuptools.package-data]
 "wrappers" = ["*.html"]